fix(trashbin): restrict retore and delete on the frontend

+++ src/components/AppNavigation/CalendarList/Trashbin.vue
@@ -70,11 +70,11 @@
 									</td>
 									<td>
 										<div class="item-actions">
-											<NcButton variant="secondary" @click="restore(item)">
+											<NcButton v-if="item.canRestore" variant="secondary" @click="restore(item)">
 												{{ t('calendar', 'Restore') }}
 											</NcButton>
 
-											<NcActions :forceMenu="true">
+											<NcActions v-if="item.canDelete" :forceMenu="item.canRestore">
 												<NcActionButton @click="onDeletePermanently(item)">
 													<template #icon>
 														<IconDelete :size="20" decorative />
@@ -164,6 +164,7 @@ export default {
 				url: calendar._url,
 				deletedAt: calendar._props['{http://nextcloud.com/ns}deleted-at'],
 				color: calendar.color ?? uidToHexColor(calendar.displayname),
+				readonly: false,
 			}))
 			const formattedCalendarObjects = this.objects.map((vobject) => {
 				let eventSummary = t('calendar', 'Untitled item')
@@ -185,6 +186,8 @@ export default {
 				const color = vobject.calendarComponent.getComponentIterator().next().value?.color
 					?? vobject.calendar?.color
 					?? uidToHexColor(subline)
+				const canDelete = vobject.calendar?.canDeleteObject
+				const canRestore = vobject.calendar?.canCreateObject && vobject.calendar?.canModifyObject
 				return {
 					vobject,
 					type: 'object',
@@ -194,6 +197,8 @@ export default {
 					url: vobject.uri,
 					deletedAt: vobject.dav._props['{http://nextcloud.com/ns}deleted-at'],
 					color,
+					canDelete,
+					canRestore,
 				}
 			})
 
@@ -281,6 +286,9 @@ export default {
 				return
 			}
 			this.items.forEach((item) => {
+				if (!this.canModify(item)) {
+					return
+				}
 				this.onDeletePermanently(item)
 			})
 		},
